refactor: rename and relocate menu-bar.tsx to menu-bar folder
